requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.5.3
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.13.0
opencv-python==4.8.0.76
Pillow==10.0.0
requests==2.31.0
twilio==8.5.0
sendgrid==6.10.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
schedule==1.2.0
rasterio==1.3.8
geopandas==0.13.2
folium==0.14.0
plotly==5.15.0
dash==2.13.0
celery==5.3.1
redis==4.6.0
reportlab==4.0.4
# LIDAR and 3D Point Cloud Processing
open3d==0.17.0
laspy==2.4.1
plyfile==1.0.1
# Deep Learning for 3D Data
torch==2.0.1
torchvision==0.15.2
pytorch3d==0.7.4
# Additional 3D Processing
trimesh==3.22.4
pyntcloud==0.3.1
# Geometric Analysis
shapely==2.0.1
rtree==1.0.1